envs/dev/flask-app/src/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
azure-cosmos==4.6.0
pydantic==2.5.3